Move line state and type into contracts
