+++ packages/contracts/src/Line/LineType.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Packages\Models\Enum;
+namespace Packages\Contracts\Line;
 
 enum LineType: string
 {
