+++ packages/models/src/Model/Line/AbstractLine.php
@@ -2,7 +2,7 @@
 
 namespace Packages\Models\Model\Line;
 
-use Packages\Models\Enum\LineType;
+use Packages\Contracts\Line\LineType;
 use Symfony\Component\Serializer\Annotation\DiscriminatorMap;
 use Symfony\Component\Serializer\Annotation\Ignore;
