+++ packages/models/src/Model/Line/Branch.php
@@ -2,7 +2,7 @@
 
 namespace Packages\Models\Model\Line;
 
-use Packages\Models\Enum\LineType;
+use Packages\Contracts\Line\LineType;
 
 class Branch extends AbstractLine
 {
