+++ services/analyse/tests/Query/LineCoverageQueryTest.php
@@ -14,8 +14,8 @@ use App\Tests\Mock\Factory\MockEnvironmentServiceFactory;
 use Google\Cloud\BigQuery\QueryResults;
 use Override;
 use Packages\Contracts\Environment\Environment;
+use Packages\Contracts\Line\LineState;
 use Packages\Contracts\Provider\Provider;
-use Packages\Models\Enum\LineState;
 use Packages\Contracts\Tag\Tag;
 use PHPUnit\Framework\Attributes\DataProvider;
 use Symfony\Component\Serializer\SerializerInterface;
