+++ services/analyse/tests/Service/CoverageAnalyserServiceTest.php
@@ -22,8 +22,8 @@ use App\Service\CoverageAnalyserService;
 use App\Service\Diff\DiffParserService;
 use App\Service\History\CommitHistoryServiceInterface;
 use App\Service\QueryService;
+use Packages\Contracts\Line\LineState;
 use Packages\Contracts\Provider\Provider;
-use Packages\Models\Enum\LineState;
 use Packages\Contracts\Tag\Tag;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
