+++ services/analyse/tests/Service/LineGroupingServiceTest.php
@@ -5,12 +5,12 @@ namespace App\Tests\Service;
 use App\Query\Result\LineCoverageQueryResult;
 use App\Service\LineGroupingService;
 use DateTimeImmutable;
+use Packages\Contracts\Line\LineState;
 use Packages\Contracts\Provider\Provider;
 use Packages\Event\Model\EventInterface;
 use Packages\Event\Model\UploadsFinalised;
 use Packages\Message\PublishableMessage\PublishableMissingCoverageAnnotationMessage;
 use Packages\Message\PublishableMessage\PublishablePartialBranchAnnotationMessage;
-use Packages\Models\Enum\LineState;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
 use Psr\Log\NullLogger;
