+++ services/ingest/tests/Service/BigQueryMetadataBuilderServiceTest.php
@@ -5,17 +5,17 @@ namespace App\Tests\Service;
 use App\Service\BigQueryMetadataBuilderService;
 use App\Tests\Mock\Factory\MockNormalizerFactory;
 use DateTimeImmutable;
+use Packages\Contracts\Line\LineType;
 use Packages\Contracts\Provider\Provider;
+use Packages\Contracts\Tag\Tag;
 use Packages\Event\Model\Upload;
 use Packages\Models\Enum\CoverageFormat;
-use Packages\Models\Enum\LineType;
 use Packages\Models\Model\Coverage;
 use Packages\Models\Model\File;
 use Packages\Models\Model\Line\AbstractLine;
 use Packages\Models\Model\Line\Branch;
 use Packages\Models\Model\Line\Method;
 use Packages\Models\Model\Line\Statement;
-use Packages\Contracts\Tag\Tag;
 use PHPUnit\Framework\Attributes\DataProvider;
 use Psr\Log\NullLogger;
 use Symfony\Bundle\FrameworkBundle\Test\KernelTestCase;
