+++ services/ingest/tests/Service/Persist/S3PersistServiceTest.php
@@ -8,17 +8,17 @@ use AsyncAws\SimpleS3\SimpleS3Client;
 use DateTimeImmutable;
 use Packages\Contracts\Environment\Environment;
 use Packages\Contracts\Event\Event;
+use Packages\Contracts\Line\LineType;
 use Packages\Contracts\Provider\Provider;
+use Packages\Contracts\Tag\Tag;
 use Packages\Event\Model\Upload;
 use Packages\Models\Enum\CoverageFormat;
-use Packages\Models\Enum\LineType;
 use Packages\Models\Model\Coverage;
 use Packages\Models\Model\File;
 use Packages\Models\Model\Line\AbstractLine;
 use Packages\Models\Model\Line\Branch;
 use Packages\Models\Model\Line\Method;
 use Packages\Models\Model\Line\Statement;
-use Packages\Contracts\Tag\Tag;
 use PHPUnit\Framework\Attributes\DataProvider;
 use Psr\Log\NullLogger;
 use Ramsey\Uuid\Uuid;
